+++ src/constants/type.ts
@@ -1,0 +1,9 @@
+export interface PayCodeOptionsType {
+    value: number | string;
+    label: string;
+}
+
+export interface ModelCodeOptionsType {
+    value: string;
+    label: string;
+}
